modify helper to concern and some details
