+++ app/controllers/application_controller.rb
@@ -1,3 +1,6 @@
 class ApplicationController < ActionController::Base
-  include SessionsHelper
+  include Logable
+
+  helper_method :current_user, :sign_in?
+  # 把從模組拿到的這兩個方法給 view 使用
 end
