+++ app/controllers/concerns/Logable.rb
@@ -1,0 +1,29 @@
+module Logable
+  extend ActiveSupport::Concern
+  # included do  
+  # end
+  # module ClassMethods 
+  # end
+
+  def sign_in(user)
+    cookies.permanent[:remember_token] = user.remember_token 
+    self.current_user = user
+  end
+
+  def signed_in? 
+    !current_user.nil?
+  end
+
+  def current_user=(user)
+    @current_user = user 
+  end
+  
+  def current_user
+    @current_user ||= User.find_by(remember_token: cookies[:remember_token])
+  end  
+
+  def sign_out
+    self.current_user = nil
+    cookies.delete(:remember_token)
+  end
+end
