+++ app/views/users/show.html.erb
@@ -1,7 +1,7 @@
 <h1>個人資料</h1>
 <div>
-  名字：<%= @user.name%>
+  名字：<%= @user.name %>
 </div>
 <div>
-  信箱：<%= @user.email%>
+  信箱：<%= @user.email %>
 </div>
